requirements.txt
ollama
pypdf
faiss-cpu
streamlit
